Added markdown filter
