+++ CHANGES
@@ -16,6 +16,7 @@ Next feature release
   specific page.
 - Make pagination next_page and prev_page be None on the edges.
 - Allow plugins to provide publishers.
+- Added `|markdown` filter.
 
 1.1
 ---
